requirements.txt
# Core framework
farm-haystack==1.20.0

# Dense embeddings
sentence-transformers==2.2.2

# PyTorch backend
torch>=1.13.0

# Vector search (CPU)
faiss-cpu==1.7.4

# Utilities
numpy==1.23.5
pandas>=1.5
tqdm>=4.65
scikit-learn==1.1.3

sqlalchemy==1.4.49
transformers==4.30.2
